replace old wires
