+++ mrmustard/physics/wires.py
@@ -1,28 +1,81 @@
-# Copyright 2023 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""``Wires`` class for supporting tensor network functionalities."""
-
 from __future__ import annotations
-from functools import cached_property
-import numpy as np
-
+from dataclasses import dataclass, field
+from typing import Sequence
+from random import randint
+from copy import deepcopy
+from enum import Enum, auto
 from IPython.display import display
+from functools import lru_cache, cached_property
 
 from mrmustard import widgets
 
 __all__ = ["Wires"]
+
+
+class Repr(Enum):
+    UNSPECIFIED = auto()
+    BARGMANN = auto()
+    FOCK = auto()
+    QUADRATURE = auto()
+    PHASESPACE = auto()
+    CHARACTERISTIC = auto()
+
+
+class WiresType(Enum):
+    DM_LIKE = auto()  # only output ket and bra on same modes
+    KET_LIKE = auto()  # only output ket
+    UNITARY_LIKE = auto()  # such that can map ket to ket
+    CHANNEL_LIKE = auto()  # such that can map dm to dm
+    PROJ_MEAS_LIKE = auto()  # only input ket
+    POVM_LIKE = auto()  # only input ket and input bra on same modes
+    CLASSICAL_LIKE = auto()  # only classical wires
+
+
+@dataclass(slots=True)
+class QuantumWire:
+    mode: int
+    is_out: bool
+    is_ket: bool
+    index: int
+    repr: Repr = Repr.UNSPECIFIED
+    id: int = field(default_factory=lambda: randint(0, 2**32 - 1))
+
+    @property
+    def is_dv(self) -> bool:
+        return self.repr == Repr.FOCK
+
+    def __hash__(self) -> int:
+        return hash((self.mode, self.is_out, self.is_ket))
+
+    def __repr__(self) -> str:
+        return f"QuantumWire(mode={self.mode}, out={self.is_out}, ket={self.is_ket}, repr={self.repr}, index={self.index})"
+
+    def __eq__(self, other: QuantumWire) -> bool:
+        return (
+            self.mode == other.mode and self.is_out == other.is_out and self.is_ket == other.is_ket
+        )
+
+
+@dataclass(slots=True)
+class ClassicalWire:
+    mode: int
+    is_out: bool
+    index: int
+    repr: Repr = Repr.UNSPECIFIED
+    id: int = field(default_factory=lambda: randint(0, 2**32 - 1))
+
+    @property
+    def is_dv(self) -> bool:
+        return self.repr == Repr.FOCK
+
+    def __hash__(self) -> int:
+        return hash((self.mode, self.is_out, self.is_dv))
+
+    def __repr__(self) -> str:
+        return f"ClassicalWire(mode={self.mode}, out={self.is_out}, repr={self.repr}, index={self.index})"
+
+    def __eq__(self, other: ClassicalWire) -> bool:
+        return self.mode == other.mode and self.is_out == other.is_out
 
 
 class Wires:
@@ -150,7 +203,7 @@ class Wires:
         ║       ║─────▶║       ║─────▶ 0
         ╚═══════╝      ╚═══════╝
 
-    The permutation that takes the contracted representations to the standard order is also returned.
+    The permutations that standardize the CV and DV variables of the contracted reprs are also returned.
 
     Args:
         modes_out_bra: The output modes on the bra side.
@@ -159,307 +212,305 @@ class Wires:
         modes_in_ket: The input modes on the ket side.
         classical_out: The output modes for classical information.
         classical_in: The input modes for classical information.
+
+    Returns:
+        A ``Wires`` object, and the permutations that standardize the CV and DV variables.
     """
 
     def __init__(
         self,
-        modes_out_bra: set[int] | None = None,
-        modes_in_bra: set[int] | None = None,
-        modes_out_ket: set[int] | None = None,
-        modes_in_ket: set[int] | None = None,
-        classical_out: set[int] | None = None,
-        classical_in: set[int] | None = None,
-    ) -> None:
-        self.args: tuple[set, ...] = (
-            modes_out_bra or set(),
-            modes_in_bra or set(),
-            modes_out_ket or set(),
-            modes_in_ket or set(),
-            classical_out or set(),
-            classical_in or set(),
-        )
-        self._len = None
+        modes_out_bra: Sequence[int] = (),
+        modes_in_bra: Sequence[int] = (),
+        modes_out_ket: Sequence[int] = (),
+        modes_in_ket: Sequence[int] = (),
+        classical_out: Sequence[int] = (),
+        classical_in: Sequence[int] = (),
+    ):
+        self.quantum_wires = set()
+        self.classical_wires = set()
 
-        # The "parent" wires object, if any. This is ``None`` for freshly initialized
-        # wires objects, and ``not None`` for subsets.
-        self._original = None
+        for i, m in enumerate(sorted(modes_out_bra)):
+            self.quantum_wires.add(QuantumWire(mode=m, is_out=True, is_ket=False, index=i))
+        n = len(modes_out_bra)
+        for i, m in enumerate(sorted(modes_in_bra)):
+            self.quantum_wires.add(QuantumWire(mode=m, is_out=False, is_ket=False, index=n + i))
+        n += len(modes_in_bra)
+        for i, m in enumerate(sorted(modes_out_ket)):
+            self.quantum_wires.add(QuantumWire(mode=m, is_out=True, is_ket=True, index=n + i))
+        n += len(modes_out_ket)
+        for i, m in enumerate(sorted(modes_in_ket)):
+            self.quantum_wires.add(QuantumWire(mode=m, is_out=False, is_ket=True, index=n + i))
+        n += len(modes_in_ket)
+        for i, m in enumerate(sorted(classical_out)):
+            self.classical_wires.add(ClassicalWire(mode=m, is_out=True, index=n + i))
+        n += len(classical_out)
+        for i, m in enumerate(sorted(classical_in)):
+            self.classical_wires.add(ClassicalWire(mode=m, is_out=False, index=n + i))
 
-        # Adds elements to the cache when calling ``__getitem__``
-        self._mode_cache = {}
+    def copy(self) -> Wires:
+        return deepcopy(self)
 
-    @cached_property
+    ###### TRANSFORMATIONS ######
+
+    @property
     def adjoint(self) -> Wires:
         r"""
-        New ``Wires`` object obtained by swapping ket and bra wires.
+        New ``Wires`` object with the adjoint quantum wires (ket becomes bra and vice versa).
         """
-        return Wires(
-            self.args[2],
-            self.args[3],
-            self.args[0],
-            self.args[1],
-            self.args[4],
-            self.args[5],
-        )
+        w = self.copy()
+        for q in w.quantum_wires:
+            q.is_ket = not q.is_ket
+        return w
 
-    @cached_property
-    def bra(self) -> Wires:
+    @property
+    def dual(self) -> Wires:
         r"""
-        New ``Wires`` object with only bra wires.
+        New ``Wires`` object with dual quantum and classical wires (input becomes output and vice versa).
         """
-        ret = Wires(modes_out_bra=self.args[0], modes_in_bra=self.args[1])
-        ret._original = self.original or self
-        return ret
+        w = self.copy()
+        for q in w.quantum_wires:
+            q.is_out = not q.is_out
+        for c in w.classical_wires:
+            c.is_out = not c.is_out
+        return w
+
+    ###### SUBSETS ######
+
+    @lru_cache
+    def __getitem__(self, modes: tuple[int, ...] | int) -> Wires:
+        """
+        Returns the quantum and classical wires with the given modes.
+        """
+        modes = {modes} if isinstance(modes, int) else set(modes)
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires if q.mode in modes}
+        w.classical_wires = {c for c in self.classical_wires if c.mode in modes}
+        return w
 
     @cached_property
     def classical(self) -> Wires:
         r"""
         New ``Wires`` object with only classical wires.
         """
-        ret = Wires(classical_out=self.args[4], classical_in=self.args[5])
-        ret._original = self.original or self
-        return ret
+        w = Wires()
+        w.classical_wires = self.classical_wires
+        return w
 
     @cached_property
     def quantum(self) -> Wires:
         r"""
         New ``Wires`` object with only quantum wires.
         """
-        ret = Wires(
-            modes_out_bra=self.args[0],
-            modes_in_bra=self.args[1],
-            modes_out_ket=self.args[2],
-            modes_in_ket=self.args[3],
-        )
-        ret._original = self.original or self
-        return ret
+        w = Wires()
+        w.quantum_wires = self.quantum_wires
+        return w
 
     @cached_property
-    def dual(self) -> Wires:
+    def bra(self) -> Wires:
         r"""
-        New ``Wires`` object obtained by swapping input and output wires.
+        New ``Wires`` object with only quantum bra wires.
         """
-        return Wires(
-            self.args[1],
-            self.args[0],
-            self.args[3],
-            self.args[2],
-            self.args[5],
-            self.args[4],
-        )
-
-    @cached_property
-    def id(self) -> int:
-        r"""
-        A numerical identifier for this ``Wires`` object.
-
-        The ``id`` are random and unique, and are preserved when taking subsets.
-        """
-        if self.original:
-            return self.original.id
-        return np.random.randint(0, 2**31)
-
-    @cached_property
-    def ids(self) -> list[int]:
-        r"""
-        A list of numerical identifier for the wires in this ``Wires`` object, in
-        the standard order.
-
-        The ``ids`` are derived incrementally from the ``id`` and are unique.
-
-        .. code-block::
-
-            >>> w = Wires(modes_in_ket = {0,1}, modes_out_ket = {0,1})
-            >>> id = w.id
-            >>> ids = w.ids
-            >>> assert ids == [id, id+1, id+2, id+3]
-        """
-        if self.original:
-            return [self.original.ids[i] for i in self.indices]
-        return [id for d in self.ids_dicts for id in d.values()]
-
-    @cached_property
-    def ids_dicts(self) -> list[dict[int, int]]:
-        r"""
-        A list of dictionary mapping modes to ``ids``, one for each of the subsets
-        (``output.bra``, ``input.bra``, ``output.ket``, ``input.ket``,
-        ``output.classical``, and ``input.classical``).
-
-        If subsets are taken, ``ids_dicts`` refers to the parent object rather than to the
-        child.
-        """
-        if self.original:
-            return self.original.ids_dicts
-        return [{m: i + self.id for m, i in d.items()} for d in self.index_dicts]
-
-    @cached_property
-    def index_dicts(self) -> list[dict[int, int]]:
-        r"""
-        A list of dictionary mapping modes to indices, one for each of the subsets
-        (``output.bra``, ``input.bra``, ``output.ket``, ``input.ket``,
-        ``output.classical``, and ``input.classical``).
-
-        If subsets are taken, ``index_dicts`` refers to the parent object rather than to the
-        child.
-        """
-        if self.original:
-            return self.original.index_dicts
-        return [
-            {m: i + sum(len(s) for s in self.args[:t]) for i, m in enumerate(lst)}
-            for t, lst in enumerate(self.sorted_args)
-        ]
-
-    @property
-    def ids_index_dicts(self) -> list[dict[int, int]]:
-        r"""
-        A list of dictionary mapping ids to indices, one for each of the subsets
-        (``output.bra``, ``input.bra``, ``output.ket``, ``input.ket``,
-        ``output.classical``, and ``input.classical``).
-
-        If subsets are taken, ``ids_index_dicts`` refers to the parent object rather than to the
-        child.
-        """
-        if self.original:
-            return self.original.ids_index_dicts
-        return [
-            {v: self.index_dicts[t][k] for k, v in self.ids_dicts[t].items()}
-            for t in (0, 1, 2, 3, 4, 5)
-        ]
-
-    @cached_property
-    def indices(self) -> tuple[int, ...]:
-        r"""
-        The array of indices of this ``Wires`` in the standard order.
-        When a subset is selected (e.g. ``.ket``), it doesn't include wires that do not belong
-        to the subset, but it still counts them because indices refer to the original modes.
-
-        .. code-block::
-
-            >>> w = Wires(modes_in_ket = {0,1}, modes_out_ket = {0,1})
-            >>> assert w.indices == (0,1,2,3)
-            >>> assert w.input.indices == (2,3)
-        """
-        return tuple(
-            self.index_dicts[t][m] for t, modes in enumerate(self.sorted_args) for m in modes
-        )
-
-    @cached_property
-    def input(self) -> Wires:
-        r"""
-        New ``Wires`` object without output wires.
-        """
-        ret = Wires(set(), self.args[1], set(), self.args[3], set(), self.args[5])
-        ret._original = self.original or self
-        return ret
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires if not q.is_ket}
+        return w
 
     @cached_property
     def ket(self) -> Wires:
         r"""
-        New ``Wires`` object with only ket wires.
+        New ``Wires`` object with only quantum ket wires.
         """
-        ret = Wires(modes_out_ket=self.args[2], modes_in_ket=self.args[3])
-        ret._original = self.original or self
-        return ret
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires if q.is_ket}
+        return w
+
+    @cached_property
+    def input(self) -> Wires:
+        r"""
+        New ``Wires`` object with only classical and quantum input wires.
+        """
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires if not q.is_out}
+        w.classical_wires = {c for c in self.classical_wires if not c.is_out}
+        return w
+
+    @cached_property
+    def output(self) -> Wires:
+        r"""
+        New ``Wires`` object with only classical and quantum output wires.
+        """
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires if q.is_out}
+        w.classical_wires = {c for c in self.classical_wires if c.is_out}
+        return w
+
+    ###### PROPERTIES ######
+
+    @cached_property
+    def id(self) -> int:
+        return randint(0, 2**32 - 1)
 
     @cached_property
     def modes(self) -> set[int]:
         r"""
         The modes spanned by the wires.
         """
-        return set.union(*self.args)
-
-    @property
-    def original(self):
-        r"""
-        The parent wire, if any.
-        """
-        return self._original
+        return {q.mode for q in self.quantum_wires} | {c.mode for c in self.classical_wires}
 
     @cached_property
-    def output(self) -> Wires:
+    def ids(self) -> tuple[int, ...]:
         r"""
-        New ``Wires`` object with only output wires.
+        The ids of the wires in standard order.
         """
-        ret = Wires(self.args[0], set(), self.args[2], set(), self.args[4], set())
-        ret._original = self.original or self
-        return ret
+        return tuple(w.id for w in self.sorted_wires)
 
     @cached_property
-    def sorted_args(self) -> tuple[list[int], ...]:
+    def indices(self) -> tuple[int, ...]:
         r"""
-        The sorted arguments. Allows to sort them only once.
+        The indices of the wires in standard order.
         """
-        return tuple(sorted(s) for s in self.args)
+        return tuple(w.index for w in self.sorted_wires)
 
-    def contracted_indices(self, other: Wires):
+    @cached_property
+    def DV_indices(self) -> tuple[int, ...]:
         r"""
-        Returns the indices being contracted between self and other when calling matmul.
-
-        Args:
-            other: another Wires object
+        The indices of the DV wires (both quantum and classical) in standard order.
         """
-        ovlp_bra, ovlp_ket = self.overlap(other)
-        idxA = self.output.bra[ovlp_bra].indices + self.output.ket[ovlp_ket].indices
-        idxB = other.input.bra[ovlp_bra].indices + other.input.ket[ovlp_ket].indices
-        return idxA, idxB
+        return tuple(q.index for q in self.DV_wires)
 
-    def overlap(self, other: Wires) -> tuple[set[int], set[int]]:
+    @cached_property
+    def CV_indices(self) -> tuple[int, ...]:
         r"""
-        Returns the modes that overlap between the two ``Wires`` objects.
-
-        Args:
-            other: Another ``Wires`` object.
+        The indices of the CV wires (both quantum and classical) in standard order.
         """
-        ovlp_ket = self.output.ket.modes & other.input.ket.modes
-        ovlp_bra = self.output.bra.modes & other.input.bra.modes
-        return ovlp_bra, ovlp_ket
+        return tuple(q.index for q in self.CV_wires)
+
+    @cached_property
+    def DV_wires(self) -> tuple[QuantumWire | ClassicalWire, ...]:
+        r"""
+        The DV wires in standard order.
+        """
+        return tuple(w for w in self.sorted_wires.copy() if w.is_dv)
+
+    @cached_property
+    def CV_wires(self) -> tuple[QuantumWire | ClassicalWire, ...]:
+        r"""
+        The CV wires in standard order.
+        """
+        return tuple(w for w in self.sorted_wires.copy() if not w.is_dv)
+
+    @cached_property
+    def args(self) -> tuple[set[int], ...]:
+        r"""
+        The arguments to pass to ``Wires`` to create the same object with fresh wires.
+        """
+        return (
+            self.bra.output.modes,
+            self.bra.input.modes,
+            self.ket.output.modes,
+            self.ket.input.modes,
+            self.classical.output.modes,
+            self.classical.input.modes,
+        )
+
+    @cached_property
+    def wires(self) -> set[QuantumWire | ClassicalWire]:
+        r"""
+        A set of all wires.
+        """
+        return {*self.quantum_wires, *self.classical_wires}
+
+    @cached_property
+    def sorted_wires(self) -> list[QuantumWire | ClassicalWire]:
+        r"""
+        A list of all wires in standard order.
+        """
+        return [
+            *sorted(self.bra.output.wires, key=lambda s: s.mode),
+            *sorted(self.bra.input.wires, key=lambda s: s.mode),
+            *sorted(self.ket.output.wires, key=lambda s: s.mode),
+            *sorted(self.ket.input.wires, key=lambda s: s.mode),
+            *sorted(self.classical.output.wires, key=lambda s: s.mode),
+            *sorted(self.classical.input.wires, key=lambda s: s.mode),
+        ]
+
+    ###### METHODS ######
+
+    def wire(self, mode: int, is_out: bool, is_ket: bool) -> QuantumWire | ClassicalWire:
+        r"""
+        Returns the wire with the given mode, ket, and output status.
+        """
+        if quantum := [
+            w
+            for w in self.quantum_wires
+            if w.mode == mode and w.is_out == is_out and w.is_ket == is_ket
+        ]:
+            return quantum[0]
+        if classical := [w for w in self.classical_wires if w.mode == mode and w.is_out == is_out]:
+            return classical[0]
+        raise ValueError(f"No wire with mode {mode}, is_out {is_out}, and is_ket {is_ket}.")
+
+    def reindex(self) -> None:
+        r"""
+        Updates the indices of the wires according to the standard order.
+        """
+        for i, w in enumerate(self.sorted_wires):
+            w.index = i
 
     def __add__(self, other: Wires) -> Wires:
         r"""
-        New ``Wires`` object that combines the wires of ``self`` and those of ``other``.
-
-        Raises:
-            ValueError: If any leftover wires would overlap.
+        New ``Wires`` object that combines the wires of self and other.
+        If there are overlapping wires (same mode, is_ket, is_out), raises a ValueError.
         """
-        new_args = []
-        for t, (m1, m2) in enumerate(zip(self.args, other.args)):
-            if m := m1 & m2:
-                raise ValueError(f"{t}-type wires overlap at mode {m}.")
-            new_args.append(m1 | m2)
-        return Wires(*new_args)
+        if ovlp_classical := self.classical_wires & other.classical_wires:
+            raise ValueError(f"Overlapping classical wires {ovlp_classical}.")
+        if ovlp_quantum := self.quantum_wires & other.quantum_wires:
+            raise ValueError(f"Overlapping quantum wires {ovlp_quantum}.")
+        w = Wires()
+        w.quantum_wires = self.quantum_wires | other.quantum_wires
+        w.classical_wires = self.classical_wires | other.classical_wires
+        w.reindex()
+        return w
+
+    def __sub__(self, other: Wires) -> Wires:
+        r"""
+        New ``Wires`` object that removes the wires of other from self, by mode.
+        Note it does not look at ket, bra, input or output: just the mode. Use with caution.
+        """
+        w = Wires()
+        w.quantum_wires = {q for q in self.quantum_wires.copy() if q.mode not in other.modes}
+        w.classical_wires = {c for c in self.classical_wires.copy() if c.mode not in other.modes}
+        w.reindex()
+        return w
 
     def __bool__(self) -> bool:
-        r"""
-        Returns ``True`` if this ``Wires`` object has any wires, ``False`` otherwise.
-        """
-        return any(self.args)
+        return bool(self.quantum_wires) or bool(self.classical_wires)
 
-    def __eq__(self, other) -> bool:
-        return self.args == other.args
+    def __hash__(self) -> int:
+        return hash(tuple(tuple(sorted(subset)) for subset in self.args))
 
-    def __getitem__(self, modes: tuple[int, ...] | int) -> Wires:
-        r"""
-        New ``Wires`` object with wires only on the given modes.
-        """
-        modes = {modes} if isinstance(modes, int) else set(modes)
-        if tuple(modes) not in self._mode_cache:
-            w = Wires(*(self.args[t] & modes for t in (0, 1, 2, 3, 4, 5)))
-            w._original = self.original or self
-            self._mode_cache[tuple(modes)] = w
-        return self._mode_cache[tuple(modes)]
+    def __eq__(self, other: Wires) -> bool:
+        return (
+            self.quantum_wires == other.quantum_wires
+            and self.classical_wires == other.classical_wires
+        )
 
     def __len__(self) -> int:
-        r"""
-        The number of wires.
-        """
-        if self._len is None:
-            self._len = sum(map(len, self.args))
-        return self._len
+        return len(self.quantum_wires) + len(self.classical_wires)
 
-    def __matmul__(self, other: Wires) -> tuple[Wires, list[int]]:
+    def __repr__(self) -> str:
+        return (
+            f"Wires(modes_out_bra={self.output.bra.modes}, "
+            f"modes_in_bra={self.input.bra.modes}, "
+            f"modes_out_ket={self.output.ket.modes}, "
+            f"modes_in_ket={self.input.ket.modes}, "
+            f"classical_out={self.output.classical.modes}, "
+            f"classical_in={self.input.classical.modes})"
+        )
+
+    def __matmul__(self, other: Wires) -> tuple[Wires, list[int], list[int]]:
         r"""
-        Returns the wires of the circuit composition of self and other without adding missing
-        adjoints. It also returns the permutation that takes the contracted representations
-        to the standard order. An exception is raised if any leftover wires would overlap.
+        Returns the ``Wires`` for the circuit component resulting from the composition of self and other.
+        Returns also the permutations of the CV and DV wires to reorder the wires to standard order.
         Consider the following example:
 
         .. code-block::
@@ -474,76 +525,47 @@ class Wires:
         .. code-block::
 
                        ╔═══════╗
-            B|(D-A)────║self @ ║────C|(A-D)
-            b|(d-a)────║ other ║────c|(a-d)
+            B+(D-A)────║self @ ║────C+(A-D)
+            b+(d-a)────║ other ║────c+(a-d)
                        ╚═══════╝
 
-        In comparison, contracting the representations rather than the wires corresponds to
-        an order where we start from juxtaposing the objects and then removing pairs of contracted
-        indices, i.e. A-D, B, C, D-A and then the same for a-d, b, c, d-a. The returned permutation
-        is the one that takes the result of multiplying representations to the standard order.
-
-        This way it is possible to write:
+        Using the permutations, it is possible to write:
 
         .. code-block::
 
             ansatz = ansatz1[idx1] @ ansatz2[idx2]  # not in standard order
-            wires, perm = wires1 @ wires2  # matmul the wires of each component
-            ansatz = ansatz.reorder(perm)  # now in standard order
+            wires, perm_CV, perm_DV = wires1 @ wires2  # matmul the wires
+            ansatz = ansatz.reorder(perm_CV, perm_DV)  # now in standard order
 
         Args:
             other: The wires of the other circuit component.
 
         Returns:
-            The wires of the circuit composition and the permutation.
-
-        Raises:
-            ValueError: If any leftover wires would overlap.
+            The wires of the circuit composition and the permutations.
         """
-        if self.original or other.original:
-            raise ValueError("Cannot contract a subset of wires.")
-        A, B, a, b, E, F = self.args
-        C, D, c, d, G, H = other.args
-        sets = (A - D, B, a - d, b, E - H, F, C, D - A, c, d - a, G, H - E)
-        if m := sets[0] & sets[6]:
-            raise ValueError(f"Output bra modes {m} overlap.")
-        if m := sets[1] & sets[7]:
-            raise ValueError(f"Input bra modes {m} overlap.")
-        if m := sets[2] & sets[8]:
-            raise ValueError(f"Output ket modes {m} overlap.")
-        if m := sets[3] & sets[9]:
-            raise ValueError(f"Input ket modes {m} overlap.")
-        if m := sets[4] & sets[10]:
-            raise ValueError(f"Output classical modes {m} overlap.")
-        if m := sets[5] & sets[11]:
-            raise ValueError(f"Input classical modes {m} overlap.")
-        bra_out = sets[0] | sets[6]  # (self.output.bra - other.input.bra) | other.output.bra
-        bra_in = sets[1] | sets[7]  # self.input.bra | (other.input.bra - self.output.bra)
-        ket_out = sets[2] | sets[8]  # (self.output.ket - other.input.ket) | other.output.ket
-        ket_in = sets[3] | sets[9]  # self.input.ket | (other.input.ket - self.output.ket)
-        classical_out = (
-            sets[4] | sets[10]
-        )  # (self.output.classical - other.input.classical) | other.output.classical
-        classical_in = (
-            sets[5] | sets[11]
-        )  # self.input.classical | (other.input.classical - self.output.classical)
-        w = Wires(bra_out, bra_in, ket_out, ket_in, classical_out, classical_in)
+        bra_out = other.output.bra + (self.output.bra - other.input.bra)
+        ket_out = other.output.ket + (self.output.ket - other.input.ket)
+        bra_in = self.input.bra + (other.input.bra - self.output.bra)
+        ket_in = self.input.ket + (other.input.ket - self.output.ket)
+        cl_out = other.classical.output + (self.classical.output - other.classical.input)
+        cl_in = self.classical.input + (other.classical.input - self.classical.output)
 
-        # preserve ids
-        for t in (0, 1, 2, 3, 4, 5):
-            for m in w.args[t]:
-                w.ids_dicts[t][m] = self.ids_dicts[t][m] if m in sets[t] else other.ids_dicts[t][m]
+        # get the wires
+        w = Wires()
+        w.quantum_wires = (bra_out + bra_in + ket_out + ket_in).wires
+        w.classical_wires = (cl_out + cl_in).wires
+        w.reindex()
 
-        # calculate permutation
-        result_ids = [id for d in w.ids_dicts for id in d.values()]
-        self_other_ids = [
-            self.ids_dicts[t][m] for t in (0, 1, 2, 3, 4, 5) for m in sorted(sets[t])
-        ] + [other.ids_dicts[t][m] for t in (0, 1, 2, 3, 4, 5) for m in sorted(sets[t + 6])]
-        perm = [self_other_ids.index(id) for id in result_ids]
-        return w, perm
-
-    def __repr__(self) -> str:
-        return f"Wires{self.args}"
+        # get the permutations
+        CV_ids = [w.id for w in w.CV_wires if w.id in self.ids] + [
+            w.id for w in w.CV_wires if w.id in other.ids
+        ]
+        DV_ids = [w.id for w in w.DV_wires if w.id in self.ids] + [
+            w.id for w in w.DV_wires if w.id in other.ids
+        ]
+        CV_perm = [CV_ids.index(w.id) for w in w.CV_wires]
+        DV_perm = [DV_ids.index(w.id) for w in w.DV_wires]
+        return w, CV_perm, DV_perm
 
     def _ipython_display_(self):
         display(widgets.wires(self))
